// tools/types/db.ts
export const Constructor = z.object({
	constructorId: z.number(),
	constructorRef: z.string(),
	name: z.string(),
	nationality: z.string().nullable(),
	url: z.string(),
});
// ...
export const Result = z.object({
	resultId: z.number(),
	raceId: z.number(),
	driverId: z.number(),
	constructorId: z.number(),
	number: z.number(),
	grid: z.number(),
	position: z.number().nullable(),
	positionText: z.string(),
	positionOrder: z.number(),
	points: z.number(),
	laps: z.number(),
	time: z.string().nullable(),
	milliseconds: z.number().nullable(),
	fastestLap: z.number().nullable(),
	rank: z.number(),
	fastestLapTime: z.string().nullable(),
	fastestLapSpeed: z.string().nullable(),
	statusId: z.number(),
});

export const Season = z.object({
	year: z.number(),
	url: z.string(),
});
// ...
export const ConstructorChampionship = z.object({
	driverChampionshipId: z.number(),
	year: z.number(),
	raceId: z.number(),
	driverId: z.number(),
	position: z.number().nullable(),
});